pyproject.toml
[project]
name = "gemini-cli3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.2",
    "openai-agents>=0.5.1",
    "uvicorn>=0.38.0",
]
